Kotlin complained about java.Long. And suggested to use jet.Long But there is no MAX_VALUE. JetBrains, please, fix it! So, I've switched to Integer.

+++ src/main/kotlin/App.kt
@@ -1,12 +1,11 @@
 package app
 
 import org.springframework.context.annotation.AnnotationConfigApplicationContext
-import java.lang.Long
 
 fun main(args : Array<String>) {
     val ctx = AnnotationConfigApplicationContext(javaClass<AppContext>(), javaClass<WebServerContext>())
     ctx.start()
     print("Running...")
     ctx.registerShutdownHook()
-    Thread.sleep(Long.MAX_VALUE)
+    Thread.sleep(Integer.MAX_VALUE.toLong()) // Kotlin complains, that there is no MAX_VALUE for jet.Long and forbid to use java.Long
 }
